third_party/blink/web_tests/external/wpt/css/css-animations/parsing/animation-delay-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['animation-delay'] = "1s 2s" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-delay'] = "1s 2s, 3s" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-delay'] = "1s, 2s 3s" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['animation-delay'] = "1s 2s" should set animation-delay-end
  assert_equals: animation-delay-end should be canonical expected (string) "2s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s" should set animation-delay-start
  assert_equals: animation-delay-start should be canonical expected (string) "1s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s" should not set unrelated longhands
  assert_true: expected true got false
[FAIL] e.style['animation-delay'] = "1s" should set animation-delay-end
  assert_equals: animation-delay-end should be canonical expected (string) "0s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s" should set animation-delay-start
  assert_equals: animation-delay-start should be canonical expected (string) "1s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s" should not set unrelated longhands
  assert_equals: expected 0 but got 1
[FAIL] e.style['animation-delay'] = "1s 2s, 3s 4s" should set animation-delay-end
  assert_equals: animation-delay-end should be canonical expected (string) "2s, 4s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s, 3s 4s" should set animation-delay-start
  assert_equals: animation-delay-start should be canonical expected (string) "1s, 3s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s, 3s 4s" should not set unrelated longhands
  assert_true: expected true got false
[FAIL] e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should set animation-delay-end
  assert_equals: animation-delay-end should be canonical expected (string) "2s, 0s, 5s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should set animation-delay-start
  assert_equals: animation-delay-start should be canonical expected (string) "1s, 3s, 4s" but got (undefined) undefined
[FAIL] e.style['animation-delay'] = "1s 2s, 3s, 4s 5s" should not set unrelated longhands
  assert_true: expected true got false
Harness: the test ran to completion.
